# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyticsSummary": {
        "additionalProperties": false,
        "properties": {
          "computedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "When the figures were last computed; null if they haven't been yet."
          },
          "definitions": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "What each figure means (design §7: every card shows its definition).",
            "type": "object"
          },
          "metrics": {
            "$ref": "#/components/schemas/Figures"
          },
          "minCohort": {
            "description": "Groups smaller than this are suppressed.",
            "type": "number"
          },
          "node": {
            "$ref": "#/components/schemas/NodeRef"
          }
        },
        "required": [
          "computedAt",
          "definitions",
          "metrics",
          "minCohort",
          "node"
        ],
        "type": "object"
      },
      "ApiErrorBody": {
        "additionalProperties": false,
        "description": "The body of every error response.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "ChildrenBreakdown": {
        "additionalProperties": false,
        "properties": {
          "average": {
            "$ref": "#/components/schemas/Figures",
            "description": "Per child: counts are the parent's total ÷ the number of children; ratios the parent's."
          },
          "children": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "computedAt": {
                  "anyOf": [
                    {
                      "format": "date-time",
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "metrics": {
                  "$ref": "#/components/schemas/Figures"
                },
                "node": {
                  "$ref": "#/components/schemas/NodeRef"
                }
              },
              "required": [
                "node",
                "computedAt",
                "metrics"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "computedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "definitions": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "What each figure means (design §7: every card shows its definition).",
            "type": "object"
          },
          "minCohort": {
            "description": "Groups smaller than this are suppressed.",
            "type": "number"
          },
          "node": {
            "$ref": "#/components/schemas/NodeRef"
          },
          "total": {
            "$ref": "#/components/schemas/Figures",
            "description": "The parent's figures, as shown next to the children."
          }
        },
        "required": [
          "average",
          "children",
          "computedAt",
          "definitions",
          "minCohort",
          "node",
          "total"
        ],
        "type": "object"
      },
...
      "CreateHouseholdDto": {
        "properties": {
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "id": {
            "description": "Generated on the phone, so later offline changes can refer to it.",
            "format": "uuid",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/LocationDto"
          },
          "pollingStationId": {
            "description": "One of the caller's booths.",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "pollingStationId",
          "address"
        ],
        "type": "object"
      },
...
      "HouseholdCreated": {
        "additionalProperties": false,
        "properties": {
          "displayAddress": {
            "type": "string"
          },
          "duplicate": {
            "description": "True when a household with this client-generated `id` already existed.",
            "type": "boolean"
          },
          "houseKey": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lastVisit": {
            "anyOf": [
              {
                "additionalProperties": false,
                "properties": {
                  "completedAt": {
                    "anyOf": [
                      {
                        "format": "date-time",
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "id": {
                    "type": "string"
                  },
                  "memberIdsMet": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "outcome": {
                    "$ref": "#/components/schemas/VisitOutcome"
                  },
                  "startedAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "volunteerId": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "outcome",
                  "startedAt",
                  "completedAt",
                  "volunteerId",
                  "memberIdsMet"
                ],
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "The latest visit that no later visit corrects."
          },
          "location": {
            "anyOf": [
              {
                "additionalProperties": false,
                "properties": {
                  "accuracyM": {
                    "type": [
                      "number",
                      "null"
                    ]
                  },
                  "capturedAt": {
                    "anyOf": [
                      {
                        "format": "date-time",
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  }
                },
                "required": [
                  "lat",
                  "lng",
                  "accuracyM",
                  "capturedAt"
                ],
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Captured with consent; null when there is none."
          },
          "members": {
            "description": "Active members, in roll order; volunteer-added members last.",
            "items": {
              "$ref": "#/components/schemas/MemberSummary"
            },
            "type": "array"
          },
          "origin": {
            "$ref": "#/components/schemas/RecordOrigin"
          },
          "partId": {
            "type": "string"
          },
          "pollingStationId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/HouseholdStatus"
          },
          "structuredAddress": {}
        },
        "required": [
          "displayAddress",
          "duplicate",
          "houseKey",
          "id",
          "lastVisit",
          "location",
          "members",
          "origin",
          "partId",
          "pollingStationId",
          "status",
          "structuredAddress"
        ],
        "type": "object"
      },
      "HouseholdDetail": {
        "additionalProperties": false,
        "properties": {
          "displayAddress": {
            "type": "string"
          },
          "houseKey": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "lastVisit": {
            "anyOf": [
              {
                "additionalProperties": false,
                "properties": {
                  "completedAt": {
                    "anyOf": [
                      {
                        "format": "date-time",
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "id": {
                    "type": "string"
                  },
                  "memberIdsMet": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "outcome": {
                    "$ref": "#/components/schemas/VisitOutcome"
                  },
                  "startedAt": {
                    "format": "date-time",
                    "type": "string"
                  },
                  "volunteerId": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "outcome",
                  "startedAt",
                  "completedAt",
                  "volunteerId",
                  "memberIdsMet"
                ],
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "The latest visit that no later visit corrects."
          },
          "location": {
            "anyOf": [
              {
                "additionalProperties": false,
                "properties": {
                  "accuracyM": {
                    "type": [
                      "number",
                      "null"
                    ]
                  },
                  "capturedAt": {
                    "anyOf": [
                      {
                        "format": "date-time",
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "lat": {
                    "type": "number"
                  },
                  "lng": {
                    "type": "number"
                  }
                },
                "required": [
                  "lat",
                  "lng",
                  "accuracyM",
                  "capturedAt"
                ],
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Captured with consent; null when there is none."
          },
          "members": {
            "description": "Active members, in roll order; volunteer-added members last.",
            "items": {
              "$ref": "#/components/schemas/MemberSummary"
            },
            "type": "array"
          },
          "origin": {
            "$ref": "#/components/schemas/RecordOrigin"
          },
          "partId": {
            "type": "string"
          },
          "pollingStationId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/HouseholdStatus"
          },
          "structuredAddress": {}
        },
        "required": [
          "id",
          "partId",
          "pollingStationId",
          "displayAddress",
          "houseKey",
          "structuredAddress",
          "location",
          "origin",
          "status",
          "members",
          "lastVisit"
        ],
        "type": "object"
      },
      "HouseholdPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HouseholdSummary"
            },
            "type": "array"
          },
          "nextCursor": {
            "description": "Pass as `cursor` to get the next page; null on the last page.",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "HouseholdUpdated": {
        "additionalProperties": false,
        "properties": {
          "household": {
            "$ref": "#/components/schemas/HouseholdDetail"
          },
          "results": {
            "additionalProperties": false,
            "description": "One result per field sent.",
            "properties": {
              "address": {
                "$ref": "#/components/schemas/FieldChangeResult"
              },
              "location": {
                "$ref": "#/components/schemas/FieldChangeResult"
              }
            },
            "type": "object"
          }
        },
        "required": [
          "household",
          "results"
        ],
        "type": "object"
      },
...
      "Revisions": {
        "additionalProperties": false,
        "properties": {
          "current": {
            "additionalProperties": false,
            "description": "Against the previous revision, summed over the parts below (design §7).",
            "properties": {
              "additions": {
                "$ref": "#/components/schemas/Figure"
              },
              "deletions": {
                "$ref": "#/components/schemas/Figure"
              },
              "net": {
                "$ref": "#/components/schemas/Figure"
              }
            },
            "required": [
              "additions",
              "deletions",
              "net"
            ],
            "type": "object"
          },
          "definitions": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "What each figure means (design §7: every card shows its definition).",
            "type": "object"
          },
          "minCohort": {
            "description": "Groups smaller than this are suppressed.",
            "type": "number"
          },
          "node": {
            "$ref": "#/components/schemas/NodeRef"
          },
          "versions": {
            "description": "A part's revisions, oldest first (empty above part level).",
            "items": {
              "additionalProperties": false,
              "properties": {
                "additions": {
                  "$ref": "#/components/schemas/Figure"
                },
                "createdAt": {
                  "format": "date-time",
                  "type": "string"
                },
                "deletions": {
                  "$ref": "#/components/schemas/Figure"
                },
                "id": {
                  "type": "string"
                },
                "revisionType": {
                  "type": "string"
                },
                "revisionYear": {
                  "type": "number"
                },
                "voters": {
                  "$ref": "#/components/schemas/Figure"
                }
              },
              "required": [
                "id",
                "revisionYear",
                "revisionType",
                "createdAt",
                "voters",
                "additions",
                "deletions"
              ],
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "current",
          "definitions",
          "minCohort",
          "node",
          "versions"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "REST API for the BoothConnect apps. Sign in with a one-time code (`/auth/otp/*`) and send the access token as `Authorization: Bearer …`. Errors are `{ requestId, code, message, details? }`.",
    "title": "BoothConnect API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/analytics/nodes/{id}/children": {
      "get": {
        "operationId": "Analytics_children",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Sort the children by this figure (e.g. `electors.total`); default: by code.",
            "in": "query",
            "name": "metric",
            "required": false,
            "schema": {
              "enum": [
                "electors.total",
                "electors.male",
                "electors.female",
                "electors.thirdGender",
                "electors.unknown",
                "ages.18-19",
                "ages.20-29",
                "ages.30-39",
                "ages.40-49",
                "ages.50-59",
                "ages.60-79",
                "ages.80+",
                "ages.unknown",
                "households.total",
                "households.large",
                "revisions.additions",
                "revisions.deletions",
                "quality.rowsExtracted",
                "quality.rowsCorrected",
                "quality.rowsRejected",
                "quality.duplicateEpics",
                "quality.missingAge",
                "quality.missingGender",
                "fieldWork.householdsAssigned",
                "fieldWork.householdsVisited",
                "fieldWork.votersMet",
                "fieldWork.outcomes.completed",
                "fieldWork.outcomes.partially_completed",
                "fieldWork.outcomes.no_one_available",
                "fieldWork.outcomes.refused",
                "fieldWork.outcomes.address_not_found",
                "fieldWork.outcomes.household_moved",
                "fieldWork.outcomes.voter_deceased",
                "fieldWork.outcomes.duplicate_or_incorrect_listing",
                "fieldWork.outcomes.follow_up_requested",
                "fieldWork.outcomes.unsafe_or_inaccessible",
                "genderRatio",
                "medianAge",
                "votersPerHousehold",
                "extractionQuality",
                "visitedShare",
                "revisions.net"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChildrenBreakdown"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/nodes/{id}/revisions": {
      "get": {
        "operationId": "Analytics_revisions",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Revisions"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Analytics"
        ]
      }
    },
    "/v1/analytics/nodes/{id}/summary": {
      "get": {
        "operationId": "Analytics_summary",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsSummary"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Analytics"
        ]
      }
    },
...
    "/v1/households": {
      "get": {
        "operationId": "Households_list",
        "parameters": [
          {
            "description": "One booth (polling station); default: all of the caller's booths.",
            "in": "query",
            "name": "boothId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "description": "Address, house number, member name or EPIC number.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "not_visited",
                "visited",
                "follow_up"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 200,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Opaque: the `nextCursor` of the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "maxLength": 512,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdPage"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Households"
        ]
      },
      "post": {
        "operationId": "Households_create",
        "parameters": [
          {
            "description": "A unique key per logical request (8–128 of A–Z a–z 0–9 _ -). Retrying with the same key returns the first response instead of repeating the change.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHouseholdDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdCreated"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Households"
        ]
      }
    },
    "/v1/households/{id}": {
      "get": {
        "operationId": "Households_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdDetail"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Households"
        ]
      },
      "patch": {
        "operationId": "Households_update",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "A unique key per logical request (8–128 of A–Z a–z 0–9 _ -). Retrying with the same key returns the first response instead of repeating the change.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHouseholdDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdUpdated"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Households"
        ]
      }
    },
    "/v1/households/{id}/members": {
      "post": {
        "operationId": "Households_addMember",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "A unique key per logical request (8–128 of A–Z a–z 0–9 _ -). Retrying with the same key returns the first response instead of repeating the change.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMemberDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberCreated"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Households"
        ]
      }
    },
...
  },
  "servers": [],
...
}